# openapi/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "SchoolOS API Documentation",
        "version": "0.1.0",
        "description": "SchoolOS foundation API contract. Product-domain routes remain provisional until their owning slices are approved."
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1"
        }
    ],
    "paths": {
...
        "/auth/verification/request": {
            "post": {
                "operationId": "v1.auth.verification.request",
                "tags": [
                    "ContactVerification"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RequestContactVerificationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "integer",
                                    "const": 202
                                }
                            }
                        },
                        "headers": {
                            "X-Request-ID": {
                                "description": "Request correlation identifier.",
                                "required": true,
                                "schema": {
                                    "type": "string",
                                    "format": "uuid"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "The request failed validation.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/components/schemas/ApiError"
                                        }
                                    },
                                    "required": [
                                        "error"
                                    ]
                                }
                            }
                        },
                        "headers": {
                            "X-Request-ID": {
                                "description": "Request correlation identifier.",
                                "required": true,
                                "schema": {
                                    "type": "string",
                                    "format": "uuid"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication is required or the supplied credentials are invalid.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/components/schemas/ApiError"
                                        }
                                    },
                                    "required": [
                                        "error"
                                    ]
                                }
                            }
                        },
                        "headers": {
                            "X-Request-ID": {
                                "description": "Request correlation identifier.",
                                "required": true,
                                "schema": {
                                    "type": "string",
                                    "format": "uuid"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "The authenticated actor is not authorized for this operation.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/components/schemas/ApiError"
                                        }
                                    },
                                    "required": [
                                        "error"
                                    ]
                                }
                            }
                        },
                        "headers": {
                            "X-Request-ID": {
                                "description": "Request correlation identifier.",
                                "required": true,
                                "schema": {
                                    "type": "string",
                                    "format": "uuid"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The requested resource was not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/components/schemas/ApiError"
                                        }
                                    },
                                    "required": [
                                        "error"
                                    ]
                                }
                            }
                        },
                        "headers": {
                            "X-Request-ID": {
                                "description": "Request correlation identifier.",
                                "required": true,
                                "schema": {
                                    "type": "string",
                                    "format": "uuid"
                                }
                            }
                        }
                    },
                    "405": {
                        "description": "The HTTP method is not supported for this route.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/components/schemas/ApiError"
                                        }
                                    },
                                    "required": [
                                        "error"
                                    ]
                                }
                            }
                        },
                        "headers": {
                            "X-Request-ID": {
                                "description": "Request correlation identifier.",
                                "required": true,
                                "schema": {
                                    "type": "string",
                                    "format": "uuid"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "The server could not complete the request.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/components/schemas/ApiError"
                                        }
                                    },
                                    "required": [
                                        "error"
                                    ]
                                }
                            }
                        },
                        "headers": {
                            "X-Request-ID": {
                                "description": "Request correlation identifier.",
                                "required": true,
                                "schema": {
                                    "type": "string",
                                    "format": "uuid"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "The service is temporarily unavailable.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/components/schemas/ApiError"
                                        }
                                    },
                                    "required": [
                                        "error"
                                    ]
                                }
                            }
                        },
                        "headers": {
                            "X-Request-ID": {
                                "description": "Request correlation identifier.",
                                "required": true,
                                "schema": {
                                    "type": "string",
                                    "format": "uuid"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/verification/confirm": {
            "post": {
                "operationId": "v1.auth.verification.confirm",
                "tags": [
                    "ContactVerification"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ConfirmContactVerificationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "verified": {
                                                    "type": "boolean"
                                                }
                                            },
                                            "required": [
                                                "verified"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        },
                        "headers": {
                            "X-Request-ID": {
                                "description": "Request correlation identifier.",
                                "required": true,
                                "schema": {
                                    "type": "string",
                                    "format": "uuid"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "The request failed validation.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/components/schemas/ApiError"
                                        }
                                    },
                                    "required": [
                                        "error"
                                    ]
                                }
                            }
                        },
                        "headers": {
                            "X-Request-ID": {
                                "description": "Request correlation identifier.",
                                "required": true,
                                "schema": {
                                    "type": "string",
                                    "format": "uuid"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication is required or the supplied credentials are invalid.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/components/schemas/ApiError"
                                        }
                                    },
                                    "required": [
                                        "error"
                                    ]
                                }
                            }
                        },
                        "headers": {
                            "X-Request-ID": {
                                "description": "Request correlation identifier.",
                                "required": true,
                                "schema": {
                                    "type": "string",
                                    "format": "uuid"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "The authenticated actor is not authorized for this operation.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/components/schemas/ApiError"
                                        }
                                    },
                                    "required": [
                                        "error"
                                    ]
                                }
                            }
                        },
                        "headers": {
                            "X-Request-ID": {
                                "description": "Request correlation identifier.",
                                "required": true,
                                "schema": {
                                    "type": "string",
                                    "format": "uuid"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The requested resource was not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/components/schemas/ApiError"
                                        }
                                    },
                                    "required": [
                                        "error"
                                    ]
                                }
                            }
                        },
                        "headers": {
                            "X-Request-ID": {
                                "description": "Request correlation identifier.",
                                "required": true,
                                "schema": {
                                    "type": "string",
                                    "format": "uuid"
                                }
                            }
                        }
                    },
                    "405": {
                        "description": "The HTTP method is not supported for this route.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/components/schemas/ApiError"
                                        }
                                    },
                                    "required": [
                                        "error"
                                    ]
                                }
                            }
                        },
                        "headers": {
                            "X-Request-ID": {
                                "description": "Request correlation identifier.",
                                "required": true,
                                "schema": {
                                    "type": "string",
                                    "format": "uuid"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "The server could not complete the request.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/components/schemas/ApiError"
                                        }
                                    },
                                    "required": [
                                        "error"
                                    ]
                                }
                            }
                        },
                        "headers": {
                            "X-Request-ID": {
                                "description": "Request correlation identifier.",
                                "required": true,
                                "schema": {
                                    "type": "string",
                                    "format": "uuid"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "The service is temporarily unavailable.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/components/schemas/ApiError"
                                        }
                                    },
                                    "required": [
                                        "error"
                                    ]
                                }
                            }
                        },
                        "headers": {
                            "X-Request-ID": {
                                "description": "Request correlation identifier.",
                                "required": true,
                                "schema": {
                                    "type": "string",
                                    "format": "uuid"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT access-token authentication using the approved RS256 issuer and rotating key IDs.",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
        "schemas": {
            "ApiError": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    },
                    "details": {
                        "type": "object"
                    },
                    "request_id": {
                        "type": "string",
                        "format": "uuid"
                    }
                },
                "required": [
                    "code",
                    "message",
                    "details",
                    "request_id"
                ],
                "title": "ApiError"
            },
...
            "ConfirmContactVerificationRequest": {
                "type": "object",
                "properties": {
                    "contact": {
                        "type": "string",
                        "maxLength": 320
                    },
                    "code": {
                        "type": "string",
                        "pattern": "^[0-9]{6}$",
                        "minLength": 6,
                        "maxLength": 6
                    }
                },
                "required": [
                    "contact",
                    "code"
                ],
                "title": "ConfirmContactVerificationRequest"
            },
...
            "RequestContactVerificationRequest": {
                "type": "object",
                "properties": {
                    "contact": {
                        "type": "string",
                        "maxLength": 320
                    }
                },
                "required": [
                    "contact"
                ],
                "title": "RequestContactVerificationRequest"
            },
...
        },
...
    }
}